marvin

+++ application/controllers/backend/Calendar.php
@@ -13,8 +13,8 @@ class Calendar extends TF_Controller
 
         if ($_POST) {
             $this->load->helper('event');
-            $booking_id = (int)$this->input->get_post('booking_id');
-            $event_id = (int)$this->input->get_post('event_id');
+            $booking_id = $this->input->get_post('booking_id') ? $this->input->get_post('booking_id') : NULL;
+            $event_id = $this->input->get_post('event_id') ? $this->input->get_post('event_id') : NULL;
             $item_id = $this->input->get_post('item_id');
             $end_date = $this->input->get_post('end_date');
             $end_time = $this->input->get_post('end_time');
@@ -24,13 +24,13 @@ class Calendar extends TF_Controller
             $notes = $this->input->get_post('notes');
             $booking_item_id = (int)$this->input->get_post('booking_item_id');
             $event_title = $this->input->get_post('event_title') ? $this->input->get_post('event_title') : '';
-            $included = (int)$this->input->get_post('included');
-            $not_included = (int)$this->input->get_post('not_included');
+            $included = (int)$this->input->get_post('incl');
+            $not_included = (int)$this->input->get_post('not_incl');
             $foc = (int)$this->input->get_post('foc');
 
             $update_calendar_views = $this->input->get_post('update_calendar_views');
 
-            if ($booking_id === 0) {
+            if (!$booking_id) {
                 // create the guest.
                 $first_name = $this->input->get_post('first_name');
                 $last_name = $this->input->get_post('last_name');
@@ -65,22 +65,31 @@ class Calendar extends TF_Controller
                 'all_day' => 0,
                 'status' => $status,
                 'event_title' => $event_title,
-                'called_by' => 0,
-                'cancelled_by' => 0,
+                'called_by' => NULL,
+                'cancelled_by' => NULL,
                 'cancelled_reason' => '',
                 'date_cancelled' => 0,
                 'notes' => $notes,
+                'incl' => $included,
+                'not_incl' => $not_included,
+                'foc' => $foc,
+                'incl_os_done_number' => $this->input->get_post('incl_os_done_number'),
+                'incl_os_done_amount' => $this->input->get_post('incl_os_done_amount'),
+                'foc_os_done_number' => $this->input->get_post('foc_os_done_number'),
+                'foc_os_done_amount' => $this->input->get_post('foc_os_done_amount'),
+                'not_incl_os_done_number' => $this->input->get_post('not_incl_os_done_number'),
+                'not_incl_os_done_amount' => $this->input->get_post('not_incl_os_done_amount'),
             );
 
             if ($status === 'cancelled') {
-                $data['called_by'] = (int)$this->input->get_post('called_by');
-                $data['cancelled_by'] = (int)$this->input->get_post('cancelled_by');
+                $data['called_by'] = $this->input->get_post('called_by') ? $this->input->get_post('called_by') : NULL;
+                $data['cancelled_by'] = $this->input->get_post('cancelled_by') ? $this->input->get_post('cancelled_by') : NULL;
                 $data['cancelled_reason'] = $this->input->get_post('cancelled_reason');
                 $data['date_cancelled'] = strtotime($this->input->get_post('date_cancelled'));
             }
 
-            $data['start_dt'] = date('c', strtotime($start_date . ' ' . $start_time));
-            $data['end_dt'] = date('c', strtotime($end_date . ' ' . $end_time));
+            $data['start_dt'] = date('Y-m-d H:i:s', strtotime($start_date . ' ' . $start_time));
+            $data['end_dt'] = date('Y-m-d H:i:s', strtotime($end_date . ' ' . $end_time));
 
             if ($facility = $this->input->get_post('facility_id')) {
                 $data['facility_id'] = $facility;
@@ -98,13 +107,13 @@ class Calendar extends TF_Controller
             $params = array();
             $params['room'] = isset($data['facility_id']) ? $data['facility_id'] : false;
             $params['people'] = $assignedTo;
-            $params['start'] = date('Y-m-d H:i:s', strtotime($data['start_dt']));
-            $params['end'] = date('Y-m-d H:i:s', strtotime($data['end_dt']));
+            $params['start'] = $data['start_dt'];
+            $params['end'] = $data['end_dt'];
             $params['event'] = $event_id;
             $params['status'] = $status;
             $params['booking_id'] = $booking_id;
             $params['item_id'] = $item_id;
-            $params['exclude_status'] = array('cancelled', 'completed', 'no show');
+            $params['exclude_status'] = array('cancelled', 'completed', 'no-show');
 
             $this->load->library('availability', $params);
             if (!$this->availability->validate()) {
@@ -208,6 +217,7 @@ class Calendar extends TF_Controller
             } else {
                 $data['author_id'] = get_current_user_id();
                 $data['entry_date'] = now();
+
                 $this->db->insert('booking_events', $data);
                 $event_id = $this->db->insert_id();
                 $item_data = $this->item->get($item_id);
@@ -217,10 +227,13 @@ class Calendar extends TF_Controller
                     if ($assignedTo) {
                         $event_users = array();
                         foreach ($assignedTo as $user) {
-                            $event_users[] = array('event_id' => $event_id, 'staff_id' => $user);
+                            if ($user)
+                                $event_users[] = array('event_id' => $event_id, 'staff_id' => $user);
                         }
-                        $this->db->delete('booking_event_users', 'event_id=' . $event_id);
-                        $this->db->insert_batch('booking_event_users', $event_users);
+                        if ($event_users) {
+                            $this->db->delete('booking_event_users', 'event_id=' . $event_id);
+                            $this->db->insert_batch('booking_event_users', $event_users);
+                        }
                     }
 
                     $current_user = get_current_user_data();
@@ -308,8 +321,41 @@ class Calendar extends TF_Controller
             }
         }
 
-        $styles = get_statuses_style(2);
+
         $inline_css = array();
+        $inline_css[] = array('name' => array(
+            '.fc-event-included',
+            '.fc-event-included .fc-bg',
+            '.fc-event-included .fc-content',
+            '.fc-event-included .fc-content .fc-time',
+            '.fc-event-included .fc-content .fc-title'
+        ), 'style' => 'background-color: #FFFF00;');
+
+        $inline_css[] = array('name' => array(
+            '.fc-event-upsell',
+            '.fc-event-upsell .fc-bg',
+            '.fc-event-upsell .fc-content',
+            '.fc-event-upsell .fc-content .fc-time',
+            '.fc-event-upsell .fc-content .fc-title'
+        ), 'style' => 'background-color: #E6E6FA;');
+
+        $inline_css[] = array('name' => array(
+            '.fc-event-foc',
+            '.fc-event-foc .fc-bg',
+            '.fc-event-foc .fc-content',
+            '.fc-event-foc .fc-content .fc-time',
+            '.fc-event-foc .fc-content .fc-title'
+        ), 'style' => 'background-color: #FFA500');
+
+        $inline_css[] = array('name' => array(
+            '.fc-event-alacarte',
+            '.fc-event-alacarte .fc-bg',
+            '.fc-event-alacarte .fc-content',
+            '.fc-event-alacarte .fc-content .fc-time',
+            '.fc-event-alacarte .fc-content .fc-title'
+        ), 'style' => 'background-color: #FFA500');
+
+        $styles = get_event_status_styles();
         foreach ($styles as $status => $style) {
             $inline_css[] = array(
                 'name' => array(
@@ -321,46 +367,14 @@ class Calendar extends TF_Controller
                 'style' => $style);
         }
 
-        $inline_css[] = array('name' => array(
-            '.fc-event-included',
-            '.fc-event-included .fc-bg',
-            '.fc-event-included .fc-content',
-            '.fc-event-included .fc-content .fc-time',
-            '.fc-event-included .fc-content .fc-title'
-        ), 'style' => 'background-color: #FFFF00 !important; color: #000; color:#788288 !important;');
 
-        $inline_css[] = array('name' => array(
-            '.fc-event-upsell',
-            '.fc-event-upsell .fc-bg',
-            '.fc-event-upsell .fc-content',
-            '.fc-event-upsell .fc-content .fc-time',
-            '.fc-event-upsell .fc-content .fc-title'
-        ), 'style' => 'background-color: #E6E6FA !important; color: #000;');
-
-        $inline_css[] = array('name' => array(
-            '.fc-event-foc',
-            '.fc-event-foc .fc-bg',
-            '.fc-event-foc .fc-content',
-            '.fc-event-foc .fc-content .fc-time',
-            '.fc-event-foc .fc-content .fc-title'
-        ), 'style' => 'background-color: #FFA500 !important;');
-
-        $inline_css[] = array('name' => array(
-            '.fc-event-alacarte',
-            '.fc-event-alacarte .fc-bg',
-            '.fc-event-alacarte .fc-content',
-            '.fc-event-alacarte .fc-content .fc-time',
-            '.fc-event-alacarte .fc-content .fc-title'
-        ), 'style' => 'background-color: #FFA500 !important;');
-
-
-        $statuses = get_statuses(2);
+        $statuses = get_event_statuses();
 
         
         $default_view = $this->session->userdata('default_calendar_view');
 
         $data = array();
-        $data['statuses'] = get_statuses(2);
+        $data['statuses'] = get_event_statuses();
 		$data['locations'] = $this->session->userdata('location'); // $locations;
         $data['inline_css'] = $inline_css;
         $inline_js = array(
@@ -516,7 +530,7 @@ class Calendar extends TF_Controller
             $resources[] = $info;
         }
 
-        $styles = get_statuses_style(2);
+        $styles = get_event_status_styles();
         $inline_css = array();
         foreach ($styles as $status => $style) {
             $inline_css[] = array(
@@ -563,7 +577,7 @@ class Calendar extends TF_Controller
 
 
         $data = array();
-        $data['statuses'] = get_statuses(2);
+        $data['statuses'] = get_event_statuses();
         $data['inline_css'] = $inline_css;
         $data['inline_js'] = array(
             'resources' => $resources,
@@ -588,7 +602,7 @@ class Calendar extends TF_Controller
             'show_off_providers' => false,
             'viewFullDetails' => true, //!tf_current_user_can('edit_calendar'),
             'canChange' => true,
-            'statuses' => get_statuses(2),
+            'statuses' => get_event_statuses(),
             'selected_statuses' => $this->session->userdata('calendar_view_status') ? $this->session->userdata('calendar_view_status') : array(),
             'selected_positions' => $this->session->userdata('calendar_view_positions') ? $this->session->userdata('calendar_view_positions') : array(),
         );
@@ -665,8 +679,13 @@ class Calendar extends TF_Controller
 
     public function get_available_services()
     {
+        $item_id = isset($_REQUEST['item_id']) ? $_REQUEST['item_id'] : '';
+        echo form_dropdown('item_id', available_booking_items($_REQUEST['booking_id']), $item_id,
+            'class="form-control" ' . ($item_id ? 'readonly' : ''));
+
+        /*
         $this->output->set_content_type('application/json')
-            ->set_output(json_encode(available_booking_items($_REQUEST['booking_id'])));
+            ->set_output(json_encode(available_booking_items($_REQUEST['booking_id'])));*/
     }
     
     public function get_booking_dates() {
@@ -743,7 +762,9 @@ class Calendar extends TF_Controller
                 'title' => $contact['first_name'] . "\n" . $work_plan,
                 'position' => $position, 
                 'hidden' => true,
-                'eventClassName' => 'contact-id-' . $contact['contact_id']);
+                'eventClassName' => 'contact-id-' . $contact['contact_id'],
+                'workPlan' => get_provider_day_schedule_2($contact['contact_id'], $date)
+            );
             $resources[] = $info;
             if (!isset($positions[$position])) $contact['position'] = array();
             $positions[$position][] = $info;
@@ -829,6 +850,7 @@ class Calendar extends TF_Controller
             $this->db->from('bookings');
             $this->db->where('bookings.guest_id = ', $this->session->userdata('user_id'));
             $this->db->where('bookings.status', 'confirmed');
+            $this->db->where('bookings.is_active', 1);
             $this->db->limit(1);
             $result = $this->db->get()->row_array();
             $booking_id = $result['booking_id'];
@@ -881,7 +903,7 @@ class Calendar extends TF_Controller
 
         $events = get_events($booking_id, 0, 'contact_id', $start, $start, false, 'confirmed', $statuses, $locations);
 
-        $styles = get_statuses_style(2);
+        $styles = get_event_status_styles();
         $inline_css = array();
         foreach ($styles as $status => $style) {
             $inline_css[] = array(
